// src/types/Profile.d.ts
export interface Profile {
  about?: string;
  belifs?: string;
  birthday?: string;
  birthplace?: string;
  city?: string;
  country?: string;
  dribbble?: string;
  email?: string;
  fb?: string;
  firstName?: string;
  gender?: string;
  id?: string;
  imgUrl?: string;
  incline?: string;
  lastName?: string;
  occupation?: string;
  phoneNumber?: string;
  province?: string;
  rss?: string;
  spotify?: string;
  status?: string;
  twitter?: string;
  uid?: string;
  yourWebsite?: string;
}